Restructured FabDB client to fit new namespacing.
