+++ src/Clients/FabDb/Endpoints/Card/CardConfig.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Memuya\Fab\Clients\FabDb\Endpoints\Card;
+
+use Memuya\Fab\Clients\BaseConfig;
+
+class CardConfig extends BaseConfig
+{
+    public string $identifier;
+}
