+++ src/Clients/FabDb/Endpoints/Deck/DeckConfig.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Memuya\Fab\Clients\FabDb\Endpoints\Deck;
+
+use Memuya\Fab\Clients\BaseConfig;
+
+class DeckConfig extends BaseConfig
+{
+    public string $slug;
+}
