+++ src/Clients/FabDb/Formatters/Formatter.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Memuya\Fab\Clients\FabDb\Formatters;
+
+interface Formatter
+{
+    /**
+     * The content type of the response.
+     *
+     * @return string
+     */
+    public function getContentType(): string;
+
+    /**
+     * Format the given data.
+     *
+     * @param string $data
+     * @return mixed
+     */
+    public function format(string $data): mixed;
+}
