--- src/Endpoints/Card/CardConfig.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Memuya\Fab\Endpoints\Card;
-
-use Memuya\Fab\Endpoints\BaseConfig;
-
-class CardConfig extends BaseConfig
-{
-   public string $identifier;
-}
