--- src/Endpoints/Deck/DeckConfig.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Memuya\Fab\Endpoints\Deck;
-
-use Memuya\Fab\Endpoints\BaseConfig;
-
-class DeckConfig extends BaseConfig
-{
-   public string $slug;
-}
