--- src/Endpoints/Deck/DeckEndpoint.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Memuya\Fab\Endpoints\Deck;
-
-use Memuya\Fab\Endpoints\Endpoint;
-use Memuya\Fab\Enums\HttpMethod;
-use Memuya\Fab\Endpoints\Deck\DeckConfig;
-
-class DeckEndpoint implements Endpoint
-{
-    private DeckConfig $config;
-
-    public function __construct(DeckConfig $config)
-    {
-        $this->config = $config;
-    }
-
-    public function getRoute(): string
-    {
-        return sprintf('/decks/%s', $this->config->slug);
-    }
-
-    public function getHttpMethod(): HttpMethod
-    {
-        return HttpMethod::GET;
-    }
-}
