--- src/Formatters/Formatter.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Memuya\Fab\Formatters;
-
-interface Formatter
-{
-    /**
-     * The content type of the response.
-     *
-     * @return string
-     */
-    public function getContentType(): string;
-
-    /**
-     * Format the given data.
-     *
-     * @param string $data
-     * @return mixed
-     */
-    public function format(string $data): mixed;
-}
