--- tests/Endpoints/Cards/CardsEndpointTest.php
@@ -1,16 +1,0 @@
-<?php
-
-use PHPUnit\Framework\TestCase;
-use Memuya\Fab\Endpoints\Cards\CardsConfig;
-use Memuya\Fab\Endpoints\Cards\CardsEndpoint;
-
-final class CardsEndpointTest extends TestCase
-{
-    public function testRouteIsSetCorrectly()
-    {
-        $endpoint = new CardsEndpoint(new CardsConfig(['cost' => '2']));
-
-        $this->assertIsString($endpoint->getRoute());
-        $this->assertSame('/cards?page=1&cost=2', $endpoint->getRoute());
-    }
-}
